pyproject.toml
[tool.poetry]
name = "Digital Twin"
version = "0.1.0"
description = ""
authors = ["ahmedabdelatty <ahmed.abdelatty@musala.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "1.0.0"
langchain = "0.1.4"
langchainhub = "0.1.4"
langchain_experimental = "0.0.49"
langgraph = "^0.0.20"
langchain-openai = "^0.0.5"
streamlit = "^1.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
